# src/main/resources/META-INF/openapi.yaml
openapi: 3.0.0
info:
    description: "This is the REST 'contract' for the Order System"
    version: 1.0.0
    title: "(20F) Group Project - Order System"
    contact:
        name: "(20F) CST 8277"
        email: "normanm@algonquincollege.com"
        url: "https://brightspace.algonquincollege.com/d2l/home/279461"
servers:
    -
        url: "http://localhost:{port}/{contextRoot}/{basePath}"
        variables:
            port: {
                default: "8080"
            }
            contextRoot: {
                default: "rest-orderSystem"
            }
            basePath: {
                default: "api/v1"
            }
security:
    -
        basicAuth: []
components:
    securitySchemes:
        basicAuth:
            type: http
            scheme: basic
    parameters:
        idPathParam:
            in: path
            name: id
            required: true
            schema: {
                type: integer,
                format: int64,
                description: "a resource's id",
                example: 1
            }
    responses:
        ForbiddenError:
            description: "Access to the specified resource has been forbidden"
...
        UnauthorizedUserError:
            description: "request requires HTTP authentication for USER_ROLE"
        NotFound:
            description: "The specified resource was not found"
    schemas:
        PojoBase:
            type: object
            required: [id, version]
            properties: {
                id: {
                    description: "id of entity",
                    type: integer
                },
                version: {
                    description: "version (OptLock) of entity",
                    type: integer
                },
                createdDate: {
                    description: "creation timestamp of entity",
                    type: string,
                    format: date-time
                    },
                updatedDate: {
                    description: "last-updatd timestamp of entity",
                    type: string,
                    format: date-time
                }
            }
        Customer:
            description: "(20F) Group Project - Customer entity"
            type: object
            allOf: [{
                $ref: "#/components/schemas/PojoBase"
            }]
            properties: {
                firstName: {type: string, description: "First Name of the customer", example: John},
                lastName: {type: string, description: "Last Name of the customer", example: Doe},
                email: {type: string, description: "Email of the customer", example: John.Doe@gmail.com},
                phoneNumber: {type: string, description: "Phone # of the customer", example: '+1.613.555.1212'},
                billingAddress: {$ref: '#/components/schemas/BillingAddress'},
                shippingAddress: {$ref: '#/components/schemas/ShippingAddress'}
            }
        CustomersList:
            description: "list of customers"
            type: array
            items: {
                $ref: "#/components/schemas/Customer"
            }
        Address:
            description: "(20W) Group Project - Address entity"
            discriminator:
              propertyName: type
            required: [type]
            type: object
            allOf: [{
                $ref: "#/components/schemas/PojoBase"
            }]
            properties: {
                street: {type: string, description: "Location of Address (street)", example: "1385 Woodroffe Avenue"},
                city: {type: string, description: "Location of Address (city)", example: "Ottawa"},
                state: {type: string, description: "Location of Address (State/Province)", example: "ON"},
                country: {type: string, description: "Location of Address (country)", example: "Canada"},
                postal: {type: string, description: "Postal Code", example: "K2G 1V8"}
            }
...
        BillingAddress:
            allOf: [{
                $ref: "#/components/schemas/Address"  # BillingAddress has all properties of Address
            }]
            properties: { # extra properties only for BillingAddress
                type: {type: string, decription: "type of address", example: "B"},
                alsoShipping: {type: boolean, description: "indicates that this BillingAddress is also the customer's ShippingAddress"}
            }
        ShippingAddress:
            allOf: [{
                $ref: "#/components/schemas/Address"  # ShippingAddress has all properties of Address
            }]
            properties: { # extra properties only for BillingAddress
                type: {type: string, decription: "type of address", example: "S"},                
            }
...
        Product:
            description: "(20W) Group Project - Product entity"
            discriminator:
              propertyName: type
            required: [type]
            type: object
            allOf: [{
                $ref: "#/components/schemas/PojoBase"
            }]
            properties: {
                description: {type: string, description: "Description of Product", example: "Charging Battery"},
                serialNo: {type: string, description: "Serial Number of Product", example: "12345"},
            }
        ProductsList:
            description: "list of products"
            type: array
            items: {
                $ref: "#/components/schemas/Product"
            }            
...
paths:
...
    /customer/{id}:
        get:
            summary: "Retrieves a customer by its id"
            parameters: [{
                $ref: "#/components/parameters/idPathParam"
            }]
            responses: {
                "200": {
                    description: "Ok (Successful operation)",
                    content: {
                        application/json: {
                            schema: {
                                $ref: "#/components/schemas/Customer"
                            }
                        }
                    }
                },
                "401": {
                    $ref: "#/components/responses/UnauthorizedUserError"
                },
                "403": {
                    $ref: "#/components/responses/ForbiddenError"
                },
                "404": {
                    $ref: "#/components/responses/NotFound"
                }
            }
        put:
            summary: "Updates a customer by its id"
            parameters: [{
                $ref: "#/components/parameters/idPathParam"
            }]
            requestBody: {
                description: "Customer updated to Database",
                required: true,
                content: {
                    application/json: {
                        schema: {
                            $ref: "#/components/schemas/Customer"
                        }
                    }
                }
            }
            responses: {
                "200": {
                    description: "Ok (Successful operation)",
                    content: {
                        application/json: {
                            schema: {
                                $ref: "#/components/schemas/CustomersList"
                            }
                        }
                    }
                }
            }
        delete:
            summary: "Removes a customer by its id"
            parameters: [{
                $ref: "#/components/parameters/idPathParam"
            }]
            responses: {
                "204": {
                    description: "Ok (No content, successful operation)"
                },
                "401": {
                    $ref: "#/components/responses/UnauthorizedUserError"
                },
                "403": {
                    $ref: "#/components/responses/ForbiddenError"
                },
                "404": {
                    $ref: "#/components/responses/NotFound"
                }
            }
...
    /product/{id}:
    
        get:
            summary: "Retrieves a product by its id"
            parameters: [{
                $ref: "#/components/parameters/idPathParam"
            }]
            responses: {
                "200": {
                    description: "Ok (Successful operation)",
                    required: true,
                    content: {
                        application/json: {
                            schema: {
                                $ref: "#/components/schemas/Product"
                            }
                        }
                    }
                },
                "401": {
                    $ref: "#/components/responses/UnauthorizedUserError"
                },
                "403": {
                    $ref: "#/components/responses/ForbiddenError"
                },
                "404": {
                    $ref: "#/components/responses/NotFound"
                }
            }
            
        put:
            summary: "Updates a product by its id"
            parameters: [{
                $ref: "#/components/parameters/idPathParam"
            }]
            requestBody: {
                description: "product updated to Database",
                required: true,
                content: {
                    application/json: {
                        schema: {
                            $ref: "#/components/schemas/Product"
                        }
                    }
                }
            }
            responses: {
                "200": {
                    description: "Ok (Successful operation)",
                    content: {
                        application/json: {
                            schema: {
                                $ref: "#/components/schemas/ProductsList"
                            }
                        }
                    }
                }
            } 
                  
        delete:
            summary: "Removes a product by its id"
            parameters: [{
                $ref: "#/components/parameters/idPathParam"
            }]
            responses: {
                "204": {
                    description: "Ok (No content, successful operation)"
                },
                "401": {
                    $ref: "#/components/responses/UnauthorizedUserError"
                },
                "403": {
                    $ref: "#/components/responses/ForbiddenError"
                },
                "404": {
                    $ref: "#/components/responses/NotFound"
                }
            }            
...